@@ -7,9 +7,9 @@ name = "mkdocs-click"
 description = "An MkDocs extension to generate documentation for Click command line applications"
 readme = "README.md"
 license = "Apache-2.0"
-keywords = ["click", "datadog", "mkdocs", "mkdocs-plugin"]
+keywords = ["click", "mkdocs", "mkdocs-plugin"]
 authors = [
-    {name = "Datadog", email = "packages@datadoghq.com"},
+    { name = "Ofek Lev", email = "oss@ofek.dev" },
 ]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -19,21 +19,21 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Operating System :: OS Independent",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Documentation",
     "Topic :: Software Development :: Documentation",
     "Topic :: Text Processing :: Markup :: Markdown",
     "Typing :: Typed",
 ]
 dynamic = ["version"]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "click >=8.1",
     "markdown >=3.3",
-    #min "importlib_metadata >=4.3",
 ]
 
 [project.urls]
